requirements.txt
numpy
matplotlib
COVID19Py
telepot
requests
schedule
psycopg2-binary